feat(channels): Change chan sig to have more things be options
Transducers and their handlers are now options. This relieves some of the really bad syntax where
undefined has to be passed for some parameters (null won't work).

BREAKING CHANGE: chan() signature has changed. Second and third parameters are removed and are now
the 'transducer' and 'handler' properties of the options object.